Fix libcros build by having it rdepend on synaptics.
This is not ideal since synaptics is really a build time dependency
and not a run-time one. However, since we are using --root-deps=rdeps,
we need to make this a run-time dependency until we have a better solution.

We can't put it in hard-target-depends because synaptics depends on stuff
that won't build until hard-target-depends is installed.