Experimental version of setup_board script.
The setup_board script will setup a board sysroot. The goal
is to be able to use a toolchain for one or more sysroots and
to be able to easily setup and rm -rf the sysroots. We want to
be able to setup per-board sysroots with the proper overlay,
compiler flags, etc.

The setup_board script will build the desired toolchain if
it doesn't exist and then create a board sysroot directory
under "/build/<board_name>". It will also create an emerge
wrapper script called "emerge-<board_name>". At any time
you can "rm -rf /build/<board_name>" and start over.

The board sysroot is created with a set of /etc/make.conf
that should allow for per-board configuration options in
the future.

The setup_board script wraps the toolchain binaries in a
simple script that tries to enforce that the SYSROOT variable
is used when building.

This is mostly working but there are still a few cases of
packages that won't properly build with this setup. I'm sending
this out early so that others can help test. There is also
still some work to do to try and improve the make.conf setup.

Review URL: http://codereview.chromium.org/570053