fix xscreensaver ebuild
When I tried to do a full build just now, xscreensaver failed because it was looking for pam stuff.  I'm not sure if some other piece of state was masking this before, but I'm synced up to Ted's most recent changes that made some other stuff need pam explicitly added to RDEPEND as well.  So maybe that's it.  Anyway...this fixed my build.

Review URL: http://codereview.chromium.org/600028