Change chromeos-login ebuild to only build session_manager.
The old behavior did a make all which built the unit tests as well.
This is the only ebuild that I know of that builds unit tests and
unit tests are not currently supported on ARM.  This should eventually
change to using a FEATURE or USE flag to turn on tests I assume.

Review URL: http://codereview.chromium.org/600043