Use --board to automaticaly select board overlays.
The board overlay contains a make.conf that is sym-linked so that it
is picked up by the top level make.conf.  This can be used to set
variables for U-Boot and Kernel builds.

The board overlay also contains toolchain.conf which specifies the
toolchain to use.

Review URL: http://codereview.chromium.org/593073