Remove the need for hard-target-depends by using package.provided.
This change adds a package.provided file to the profile that will
claim that many of the host build tools (hard-host-depends) are
already installed in the board sysroot even though they aren't.
These packages are only needed to build and when cross-compiling
properly we use the files installed on the build host.

This change allows us to swith from --root-deps=rdeps, which ignores
the DEPEND lines in ebuilds to using --root-deps, which will install
DEPEND into the sysroot when building packages. This obviates the
need for the hard-target-depends ebuild. That can be removed in a
future CL.

We modify the build_image step to explicitly pass --root-deps=rdeps
because in our target image we only want run-time dependencies.

This change also adds gcc and binutils to the set of packages added
to the dynamically created package.provided to reflect the set of
packages and versions from the toolchain selected for the board.

Remove DEPEND on hard-target-depends from chromeos ebuild.

If we use normal --root-deps then we shouldn't need hard-target-depends.
If we use --root-deps=rdeps, then having it in DEPEND is useless anyway.

Review URL: http://codereview.chromium.org/615003