subversion should not be a DEPEND, it should be (and already is) a hard-host-depend
Now that we actually try to cross-compile DEPENDs, it turns out that one of subversion's deps doesn't actually cross-compile.  But, we don't need to cross-compile subversion, we just need it on the host at build time.  It's already in hard-host-depends, so we're all good.

Review URL: http://codereview.chromium.org/641004