Adds firmware for intel wifi cards
It looks like iwl5000-2.ucode is being left out of the build now. I'm not sure if it was coming in from compat-wirless or through our kernel build which has now changed.

At any rate, there is an ebuild for it and everything works as expected. However, we still need to get to the bottom of how this broke in the first place.

Review URL: http://codereview.chromium.org/651013