Chrome-bin ebuild: pull latest chrome
This version of the ebuild doesn't use portage's dist files cache,
since the zip file we download always has the same name. Also, we
don't want to have to rev the manifest for every chrome revision
(since the hash changes).

As a result, we just manually wget the latest version or specified
version of the browser in the ebuild.

Also, note that at the time of writing, the latest browser fails to
load libcros, but an earlier version 38617 seems to load it fine. As
such, it may be desirable to emerge this ebuild like so:

CHROME_BUILD=38617 emerge-x86-generic -a chromeos-chrome-bin

When CHROME_BUILD is absent, the latest build is pulled.

Thanks to djmm@ for another CL that did a lot of the work.

Review URL: http://codereview.chromium.org/650112