Use our own baselayout that pre-adds users and drops dependency on openrc.
This Chromium OS version of baselayout is intended to be installed
in both the build chroot and the targets. This normalizes the set
of users and user-ids, so that when building we can chown and still
get the same user-ids in the target sysroots and final image. The
baselayout also creates the set of directories needed by
Chromium OS.

It is based on baselayout2, which uses a rather small and simple
set of files for the base layout.

Review URL: http://codereview.chromium.org/661294