media-libs/gd: cross compilation support.
This CL should be upstreamable.

The issue is that when building, the flags for libpng12 are found by
calling libpng12-config. The host version of libpng12-config is found
in $PATH, which output /usr/lib64 as the lib path.

We change PATH to be only $SYSROOT/usr/bin, so that in our case
/build/$BOARD/usr/bin/libpng12-config is called. We also prepend
$SYSROOT to the output paths, since teh tool just gives /usr/lib, and
we need /build/$BOARD/usr/lib.

Review URL: http://codereview.chromium.org/1119002