build: Update test image dependencies for autox.
This adds a dependency on python-xlib needed by autox.py
and adds autox as a dependency for development and test
systems.  My plan is for tests to import autox from within
their run_once() methods rather than at the top of their
files; this avoids autox.py needing to be installed on the
host system (which would be difficult to do right now, as
$CHROMEOS_ROOT doesn't get defined outside of the board-
specific emerge wrappers).

BUG=none
TEST=built a test image and confirmed that autox is installed, importable, and working

Review URL: http://codereview.chromium.org/1236002