Add support for a single kernel config file
Use a single or split kernel config as specified in the board or variant
make.conf overlay. Default to the arch specific split config if an
overlay or variant does not set either CHROMEOS_KERNEL_CONFIG or
CHROMEOS_KERNEL_SPLITCONFIG. CHROMEOS_KERNEL_CONFIG is set relative
to the root of the kernel source tree.

Review URL: http://codereview.chromium.org/1325001