shflags ebuild
Create a new ebuild for shflags. This installs into /usr/lib for lack
of a better place. By not installing into /usr/lib64, our scripts can
always look in the same place. I'm also not sure it belongs in lib64
since it's not an executable, but I'm open to discussion about this.

Review URL: http://codereview.chromium.org/1535014