Add support for profiles in overlays to setup_board.
We'd like to be able to have overlays participate in the
cascading profile. For example, this will enable a board
overlay to use package.mask to mask out some packages that
it can't handle yet, use package.use to set per-package USE
flags, have package.keywords, etc.

The change creates a profile on the fly for the given board
that only contains a "parents" file. The "parents" file
references the main profile and for any overlay that has a
"profiles/targets/chromeos" directory it will add that in
overlay priority order.

Review URL: http://codereview.chromium.org/1617011