Update x86 to Xorg server 1.7.6, mesa 7.7.1, intel-video-2.11.0, etc.
This updates the graphics stack to packages required in order to support
an Ironlake gpu. We did not update to Xorg server 1.8.0 and Mesa 7.8.1
due to stability issues when using Mesa 7.8.1. This change only updates
for x86 for now. We intend to move ARM to the newer set of packages but
will wait until the per-overlay profile support is working so that there
is the option to tweak things at the board level.

This change will likely break "incremental" buildbot builds since the
Xorg drivers need to be re-build, and this does not happen when using
binary packages. Individuals may also have to emerge by hand with
rebuild (emerge-<board> -av):

xf86-video-intel
xf86-input-keyboard
xf86-input-mouse
xf86-input-synaptics

if they are not rebuilt automatically.

Keyword changes to update to xorg 1.7.6, mesa 7.7.1, and intel-video-2.11.0

Review URL: http://codereview.chromium.org/1631014