Fix issue where packages can't find the python script directory properly.
Some packages are broken and so they snippet of python they use to find
the python scripts directory fails. In that case they default to

/usr/lib/python2.6/site-packages/

which is actually the right value. For packages that execute the
correct python snippet, they end up with:

/usr/lib64/python2.6/site-packages/

which is the wrong value for our SYSROOT.

One fix is to treat this as an autoconf cache variable. Since it is
specific to Chromium OS I put it in make.conf.common-target for now
for lack of a better place. Ideas welcome.

Review URL: http://codereview.chromium.org/1551027