Fix devkit-disks-part-id crasher
devkit-disks-part-id uses "%m" in some format strings, which is a glibc-specific extension to printf (see src/part-id.c in devkit-disks source).  That's fine, if you're using glibc's printf at runtime.  The problem is that, since autoconf can't build and run little tools to check for things in a cross-compile environment, the glib (not glibc, glib) build process couldn't determine that we had an acceptable version of printf -- so it failed over to using its own, internal replacement.  This replacement does NOT support %m and, apparently, segfaults if you try to use it.

Fail.

So, this CL pre-caches some settings that will make glib understand that we have a "good printf"

Review URL: http://codereview.chromium.org/1631017