Get mesa-7.8.1 working by disabling gallium intel.
The default ebuild for mesa-7.8.1 wants to turn on the gallium intel
driver. This does not seem stable and fails completely on at least
one device. This change disables gallium for the intel integrated
graphics. It was previously disabled on mesa-7.5.1 and mesa-7.7.1 that
we've been using.

Review URL: http://codereview.chromium.org/1654016