Fix libchewing build.
Now that libchewing is imported to our repository, we can no longer depend on
the timestamps being identical to those in the tarball. For this reason, we
can't depend on the autogenerated files that were included in the tarball,
because make will try to regenerate some of them. To fix this, we rerun
./autogen.sh

This hopefully will fix the build.
TEST=emerged the libchewing package
BUG=none

Review URL: http://codereview.chromium.org/1735019