Update to xorg-server-1.8.0 and re-add intel-video patch.
This change updates to xorg-server-1.8.0 and adds back the
intel video patch to reduce boot flicker (since the required
patch to xorg-server is in 1.8.0).

The update to xorg-server-1.8.0 should also fix a reported glx
version issue that we've been seeing. Version 1.2 was being
reported (as seen with glxinfo) and the new server reports
glx version 1.4.

We also revbump the x11-drivers so that they are rebuilt
automtically for people.

Review URL: http://codereview.chromium.org/1758016