kernel: set any new Kconfig options to no
Currently, the kernel build breaks if any new Kconfig options get added.
This was intentional. However, it gets in the way when cherry-picking
upstream patches.

When we cherry-pick a patch, we want to cherry-pick verbatim and not
have to extend it also update the splitconfig.

For non-upstream commits, we'd also like to keep config changes
separate from code changes. It makes the changes easier to rebase and
it makes it easier for the changes to be cherry-picked by others.

Review URL: http://codereview.chromium.org/1859003