Discover monotonic clock using compile-time check.
When xorg-server is being cross-compiled, there is currently no way
for us to detect whether the monotonic clock is available on the
target system, because we aren't able to run a test program. Currently, in
this situation, we default to not use the monotonic clock. One problem
with this situation is that the user will be treated as idle when the date
is updated.

To fix this situation, we now use a compile-time check to detect whether the
monotonic clock is available. This check can run just fine when we are
cross-compiling.

TEST=Emerge package on arm and x86, verifying that the monotonic clock is now used
BUG=none

Review URL: http://codereview.chromium.org/2125011