mark opencryptoki as stable for our platforms, since we treat it that way anyay
Spoke to msb earlier about this, and we agreed that, for packages that
we actually pull in to the chromiumos-overlay and modify, we should
really unmask them by marking them stable in the ebuilds themselves.  So, I'm doing that -- and also fixing up the dependencies so they're moar accurate.

Review URL: http://codereview.chromium.org/2170003