Fix udev issue where temporary persistent rules are written in one place and read from another.  The upstream udev release placed dynamic persistent rules (ones that get deleted every reboot) in /dev/.udev/, but the actual udev daemon read these rules out of /dev/.udev/rules.d/.  The side effect of this is that if the static directory (/etc/udev/rules.d) is not writeable (read-only root filesystem, for example), device interfaces keep getting renamed wlan0, wlan1, wlan2, etc, if they are plugged and unplugged or if udevd restarts.
In addition to the directory being incorrect, the actual filename was incorrect as well -- the filename wes set to begin with "tmp-", which breaks the lexical sorting rules that udevd uses, so even if the file was in the right directory, the 75-persistent-net-generator.rules would fire first, and increment the device number anyway.

Note that this isn't the complete solution since in addition the /dev/.udev/rules.d directory needs to be created before udevd runs.  Otherwise when udevd tries to enable inotify on that directory, it will fail, and a later created /dev/.udev/rules.d will not be monitored.  The fix for this is available in issue 2238005 since it is in a different git repository.

Review URL: http://codereview.chromium.org/2283002