Fix build failure caused by permissions problem in git.eclass. This is a revision to the fix that was submitted by rochberg in commit 5500abb3646e9a5b34347a23d55d0db7e4fe3c8e. The idea was right - temporarily setting the umask to 002 - but it was being done in the wrong place. The umask needs to be set before running the "git clone".
This fix is a hack. The underlying problem seems to be that the first
time dbus-c++ is emerged, it runs as root with a umask of 022, and
the second time it is run (as part of build-packages), it runs as
the user, who then lacks permission to write into
/var/lib/portage/distfiles/git-src/dbus-c++. It would be better to
do something that addresses this issue in the general case.

TEST=From a chroot environment, ran the following sequence, which
reproduces the failure:
   sudo emerge --unmerge dbus-c++
   sudo rm -rf /var/lib/portage/distfiles/git-src
   sudo emerge dbus-c++
   check permissions on /var/lib/portage/distfiles/git-src/dbus-c++
   emerge-x86-generic dbus-c++
Did this sequence both before and after the fix and saw that it
failed the first time, and succeeded the second. Also saw that
the mode of dbus-c++ did not include group write permission the
first time, but did include it the second time.

Review URL: http://codereview.chromium.org/2276009