make_chroot: emerge minimum number of packages
A make_chroot failure requires that make_chroot be re-run again. make_chroot
takes a long time to run so a failure is painful. This change minimizes
the numbers of packages emerged by make_chroot to just what is required.

The remaining packages are emerged as part of build_packages.

Change-Id: I2736135f3cd9ab46a97269f288357528996ccf36

Review URL: http://codereview.chromium.org/2855001