x11: update xorg-server and xf86-video-intel for seamless transitions.
This reapplies some patches that we'd been using earlier.
Specifically, it:

- by default, makes it so that X doesn't clear the root
  window so it's possible for it to start out with the
  contents of the framebuffer
- adds a canDoBGNoneRoot field to ScrnInfoRec so that
  drivers can specify that they support·
- adds support to the Intel driver for all of this

1.7.6-xserver-bg-none-root.patch is just
meego-xserver-bg-none-root.patch, minus a hacky "|| 1" that
made the new code path get used regardless of whether the
driver claimed to support it and minus the -nr flag.  I'm
not sure if we want to delete all of the unused moblin and
meego patches floating around in xorg-server's files
directory or if someone still plans to look at them some
more.

BUG=chromium-os:3224
TEST=built and ran it

Review URL: http://codereview.chromium.org/2805011