cros_workon: clone and checkout 9999 ebuilds
Add support for cloning and checking out the
development mainline branch to a working directory
suitable for development during the emerge src_unpack step.

This functionality is intended to be invoked from the
workon script via 'ebuild-<board> ebuild-file unpack'.

This allows the git hostname and repository path to
be specified in the ebuild, make.conf or environment.
Same for the mainline development branch to checkout
and push to by default.

The developer is expected to perform 'git fetch' and
'git merge' operations in native git on their own prior
to uploading using 'git cl'.

Change-Id: Id345ddcc72683fc90ad55eb3575180c68efae85f

Review URL: http://codereview.chromium.org/2818010