verity: rev to 0.0.2 and add WITH_CHROME=x support
libchrome is now only needed when building verity
for testing purposes, otherwise it can use a lightweight,
hacked up version of libchrome's logging which is in-tree.

This change reflects that CL and only builds WITH_CHROME
when FEATURES=test is supplied.

(Is this the right way to rev builds now?)

BUG=chromium-os:327
TEST=emerged for arm-generic and x86-generic with test and without test and without test directly into the chroot

Review URL: http://codereview.chromium.org/2812027