Enable x11-terms/aterm ebuild
Enable arm and x86 build of aterm, which allows portage to pick up -r2
of this ebuild.  The -r2 revision of the ebuild adds a fix for cross
compiles of aterm that was causing permissions to be set incorrectly.
Had to modify RDEPENDS to fix a dependency on libjpeg version.

BUG=chromium-os:4531
TEST=emerged on tegra2-dev-board and x86-generic, verified permissions

Review URL: http://codereview.chromium.org/2921003