add updated mosys ebuild for version 1.0.02.54
This adds an ebuild for the latest release snapshot of mosys. The contents are
the same as the previous ebuild -- version number gets updated appropriately
based on the ebuild name via the ${PV} variable.

TESTED=Compile tested using emerge-x86-generic, and the resulting binary run
on an "AGZ" platform.

Review URL: http://codereview.chromium.org/2895011