Update fastbuild patch to be more selective about what locks it kills.
Currently, the fastbuild patch kills all of the locks on /var/db/. Some of
those locks, such as the lock occurring during the postinst step, are
actually important. To improve stability of fastbuild, it makes sense to
try to get rid of as few locks as possible.

This patch just gets rid of the giant lock in vartree.py. Hopefully, this
won't have a big effect on performance, but will improve reliability.

TEST=Ran build_packages --fast and build_image --fast. Noticed no major
     differences in speed.
BUG=none

Review URL: http://codereview.chromium.org/2959012