Fix portage to create directories atomically in fast build.
If multiple ebuild processes create the same directory, portage should not
exit with an error message. Instead we should just be okay with the fact
that another process created the directory.

This is only necessary for fast build because we eliminate the giant lock that
surrounds package merges.

TEST=Create fresh chroot and do full build from source start to finish
BUG=none

Review URL: http://codereview.chromium.org/2998002