Upgrade ibus to depend on gtk-doc regardless.
ibus requires gtk-doc to run autogen.sh, which does
not accept the --disable-gtk-doc argument. Therefore,
it always looks for gtk-doc and fails if it is not
available. To fix this, I'm adding gtk-doc as an
explicit dependency, and also adding it to
package.provided because it's really a host
dependency, not a build dependency.

This issue shows up more often in parallel builds.

TEST=Ran emerge -uDNv hard-host-depends && ./build_packages --fast from clean
     chroot.
BUG=none

Change-Id: I20ff7be7dc5c01d35341d540360d6914962f9a44

Review URL: http://codereview.chromium.org/2995003