Update setup_board directory to reflect actual location on prebuilt server.
This change just restores the prebuilts to what they were earlier today before
the builtbots broke.

History: This week, a new buildbot was setup. Earlier today, the prebuilts from
the buildbot were sync'd to the mirror server. When this happened, it looks like
the prebuilt location changed. So we need to point at the new location. This
will temporarily allow people to build on arm again.

Besides the above, the arm toolchain currently fails to build from source. That
problem should be pursued separately.

TEST=Ran ./setup_board --board=arm-generic and ./setup_board --x86-generic.
Both worked, and ran significantly faster now that we have the
prebuilts again.
BUG=none

Review URL: http://codereview.chromium.org/2805084