Fix race condition in Portage symlink creation.
Some programs will create symlinks automatically, so we have to
tolerate these links being recreated during the merge process.
Portage already handles this case, but there is a race condition
in the code where Portage fails if the symlink is created at
exactly the wrong time.

To fix this, I've updated Portage to gracefully handle this
situation if the link is pointing at the right place anyway.

Because the extra symlinks are being created by programs other
than Portage, this patch should be independent from our
fastbuild patches. It's a useful fix in any case.

Note that this bug occurs very rarely and is hard to reproduce.

TEST=Ran build_packages --fast && build_image --fast
BUG=none

Review URL: http://codereview.chromium.org/2856048