Turn off iwlagn 802.11n again
This is a workaround that should be pulled out when we can find the tools and resources to get to the bottom of the Intel 802.11n issues.  Disabling 11n isn't a good option, but I'm left with the choice of whether to workaround this problem and move on to other issues that don't yet even have a workaround vs spending time now to find and fix this issue while there are many other important ones still affecting people right now.

BUG=chromium-os:4953
TEST=restarted with fix in place

Review URL: http://codereview.chromium.org/3029004