Add more host dependencies to package.provided, speeding up build times.
The packages added to package.provided in this change should only be used on
the host. Adding these packages to package.provided should speed up build
times.

Note also that app-text/docbook-xsl-stylesheets-1.75.2 is not parallel-safe
because it calls build-docbook-catalog in a post-install script, which in turn
is not parallel-safe (it updates system-wide files in a non-parallel-safe
manner). Removing this package from our board should improve the reliability
of our parallel_emerge step for board emerges (Note that this package can still
run into problems when we emerge hard-host-depends).

TEST=Full build_packages --fast --nousepkg
BUG=none

Review URL: http://codereview.chromium.org/3060020