Update Chrome to use LKGR of libcros headers.
Per discussion with chocobo, Chrome should use the LKGR of libcros
headers. Chrome expects to build against a specific version of the
libcros headers (the LKGR), even if the version of libcros that we
have installed on the system is newer. This fixes a compile issue
on the buildbots.

To avoid compatibility problems with Chrome and libcros, the libcros
developers are careful to make sure that new versions of libcros are
always compatible with the LKGR of Chrome.

TEST=Compiled Chrome. Verified it uses LKGR of libcros headers.
BUG=none

Review URL: http://codereview.chromium.org/2873084