Move environment exports from .bashrc to .bash_profile
Environment exports for the build need to be available for both
interactive and non-interactive shells. .bashrc is configured to
exit early if it is being run in a non-interactive shell, so we
need to but the exports in .bash_profile.

From the bash man page:
 When  bash is invoked as an interactive login shell, or as a non-inter-
active shell with the --login option, it first reads and executes  com-
mands  from  the file /etc/profile, if that file exists.  After reading
that file, it looks for ~/.bash_profile, ~/.bash_login, and ~/.profile,
in  that order, and reads and executes commands from the first one that
exists and is readable.

Change-Id: Id40e264bd0f1627b90eb96116c2dac540773fd6a

Review URL: http://codereview.chromium.org/3071018
Patch from Kevin Degi <kdegi@codeaurora.org>.