Update portage to support no-env-update feature which skips env_update.
Currently, our build occasionally fails with errors about about symlinks
already existing. This usually only slows down the build, because we have
retries, but it would be good to fix the problem.

This patch enables a no-env-update feature, which skips the env_update step.
This step isn't thread-safe, and isn't necessary for when we're cross-compiling
and don't actually need to run the binaries. When we start using this step in
parallel_emerge, it will speed up our build by fixing the random failures.

TEST=Ran full build --nousepkg without env_update
BUG=chromium-os:5382

Review URL: http://codereview.chromium.org/3043054