Bump croscorefonts ebuild to 0.0.2.
Change the URI to point to our new location at http://commondatastorage.googleapis.com/chromeos-localmirror/distfiles

This will pull in croscorefonts-0.0.2.tar.gz from the distfile storage,
which includes an updated version of Arimo Regular and Arimo Bold
with better hints that works well with BCI turned on. Note that BCI
was turned on for only these two fonts last week in anticipation of this
CL.

BUG=5287
TEST=Pages with Arial or 'arial, sans-serif' renders letters like 'f'
with as tall an ascent as Arial.
TBR=anush

Review URL: http://codereview.chromium.org/3034064