setup_board: fix toolchain symlink creation
Fixes a bug where the toolchain symlinks get clobbered. If a new
toolchain is emerged (same version), it will clobber the symlinks with the
new toolchain and not bother doing the mv or ln.

Fix is to create the symlink whenver the toolchain symlinks don't
already exist.

BUG=5899
TEST=Verified all three cases work correctly.

1. Verified a fresh toolchain build.
2. Verified a toolchain upgrade.
   * Simulated by copy the .real files over the symlinks.
3. Verified a setup_board with no toolchain upgrade.

Change-Id: I629b115764d5055e1ae7bfd8edb7c05a0ad798b0

Review URL: http://codereview.chromium.org/3202001