Temporary workaround to allow people to run setup_board again.
Right now, when I try to run setup_board, gcc fails to compile because it can't
find gcc-4.3.4-patches-1.1_.tar.bz2. This results in a broken build. To
work around this, I've changed some flags in setup_board to ensure we are
always able to use the prebuilts, even if they're out of date.

I've filed a P0 bug to fix this the right way: See Bug 5931.

TEST=Ran setup_board. It works again!
BUG=chromium-os:5931

Change-Id: Icbddbaa087443dfbae694482593240c1537ab3c4

Review URL: http://codereview.chromium.org/3189016