Update gcc-4.3.4-r4 to use toolchain_hardened.
This is generally a good change -- gcc-4.3.4-r4 was originally designed to work
with toolchain_hardened, so it probably won't compile with toolchain.eclass.
What's mystifying about this change is that it seems to fix gcc-4.4.3-r5.

Right now, it's a portage mystery why this is happening. asharif is planning
to investigate this further.

BUG=chromium-os:5931
TEST=confirmed that compiling gcc-4.4.3 actually compiles 4.4.3, and not 4.3.4

Change-Id: I5a49c6995f20036bb71301d7395d36c495135f3f

Review URL: http://codereview.chromium.org/3132028