autotest: split deps into autotest-deps, fix the eclass logic appropriately
* This makes sure that each package only installs the deps,configs,profilers that
it specified as its own. Also, it pulls in stuff provided by other packages, that
was already installed into /usr/local/autotest/

* As a proof of concept, all deps from the autotest.git repo are split and prebuilt
as part of the autotest-deps package, which is then depended on. This is also needed
by any packages that would be split off the main repo, yet depend on something
from upstream. Lastly it workarounds the bug, where autotest refuses to work with
empty deps/, even though the tests don't rely on any of them.

Change-Id: I116b9483ee20a0028d7fc1fc882ddc2e916a1ee1

Review URL: http://codereview.chromium.org/3110033