Update make.conf to not spend time generating manifests.
The "digest" feature was initially enabled to fix an issue where Manifests
needed to be manually updated every time an ebuild was changed. This was
annoying, so rtc enabled the "digest" feature. Since this feature regenerates
the Manifests every time you build, it kills most of the benefit of checking
Manifests (because, even if you have a corrupted download, it's likely that
emerge will automatically save the bad digests in the Manifest file for you).

Since then, we learned about the "-strict" feature, which is simpler -- it just
disables checking of Manifests entirely, so we don't even need to generate
them in the first place.

Switching from "digest" to "-strict" reduces build times, but that isn't the
main benefit; the main benefit is that it prevents emerge from processing all
the ebuild files to generate Manifests. When emerge processes all the ebuild
files to generate Manifests, it may run into an error, causing confusing
errors where breaking an unused ebuild would break other, unrelated ebuilds.

BUG=chromium-os:5931
TEST=Full build, build_image

Change-Id: I6d8b618c3f2dea54d1fd692a9a13c96c024c65f9

Review URL: http://codereview.chromium.org/3134036