autotest: introduce more class magic to deal with dependencies
* Introduce autotest use flag, set to off by default by eclass
* Depend on autotest-deps and autotest if useflag is on
* Make sure autotest-deps don't depend on themselves
* Turn on autotest flag for autotest-tests and autotest-deps

The idea of autotest useflag is a global switch. If you do USE=-autotest,
nothing autotest related is going to be built and any dependencies pulled in
any affected package, that uses this eclass

On the other hand, if you do a USE=autotest, it will make all packages that
have tests build them, based on the TESTS= selection, or defaults, and pull
in the correct dependencies.

Packages can select override the default to enabled by specifying:
IUSE=+autotest and in fact the only two packages that we have do it, because
autotest off makes no sense for those.

TESTS= is going to be visible also if autotest is off, but will have no
effect, and there is nothing that can be done to remedy the situation.

Change-Id: Iea584af43b4f931b074d29f4dbbb51518758e9a7

Review URL: http://codereview.chromium.org/3126033