Set hintstyle to full for Arimo.
It turned out that hintstyle needs to be set to full to turn on BCI.
That is, setting hingting to true and autohint to false is not
sufficient to turn on BCI if hintstyle is slight.

BUG=6252
TEST='gmerge fontconfig' and restart ChromeOS. To see BCI display a page
with Arimo at 13px (eg, gmail) and look at the lower case 'f'. There
should be 2 pixels in the aperture (between the crossbar and the top
curve).

Review URL: http://codereview.chromium.org/3265006