X: Map Alt-Up and Alt-Down to Page Up and Page Down.
The XKB changes to do this were pretty straightforward
(largely with the help of an old mailing list thread at
http://thread.gmane.org/gmane.comp.xfree86.internationalization/1127 and
http://thread.gmane.org/gmane.comp.xfree86.internationalization/1177).

I quickly ran into an XKB bug that prevents key redirects
from working at all.  Luckily, a fix for that just got
upstreamed about two weeks ago, so I'm pulling it in here.

I also ran into a second issue mentioned on the mailing
list, where autorepeat is disabled for the keys that I
remap.  After staring at the XKB code for a long time, I
have a tiny deletion that appears to make it work without
causing any noticeable ill effects.  I'll follow up on the
X.org list to find the right way to fix this, but I'd like
to check this in in the meantime.

BUG=chromium-os:5808
TEST=tried it out

Review URL: http://codereview.chromium.org/3364001

Change-Id: I6217af46507a37c86ec8ca4f86736d403da880b1