build: Remove NDEBUG from libcros, for now.
I'm going to check in a change to make the libcros package
fail to compile if it's being compiled with an NDEBUG
setting that differs from libbase's, so I need to remove
NDEBUG here first.  I'll add a USE flag to set it globally
later.

Change-Id: I5b66c1e3cc26737f758606e5cc4ce1a12e39dff1

BUG=chromium-os:1124
TEST=built it

Review URL: http://codereview.chromium.org/3297009