Buildbots should only contact git servers during "repo sync" step.
Currently, our buildbots download the source code twice. First, the
buildbots checkout the source code during the "repo sync" step. Next,
they checkout the source code again during the build_packages step.

This "double-checkout" was originally implemented as part of
cros-workon. As I understand it, the reasoning behind the
double-checkout is that we want to pin the ebuild to a specific
revision.

In the past few weeks, problems with build_packages related to git
flakiness has lead to hard-to-debug failures, both on the buildbot and
on developer systems. See bug 5990 and bug 6453 for discussion. Bug
6198 might also be related.

Fortunately, this pinning can be accomplished without downloading the
source code twice -- we can just make a copy of our local repository and
switch to the appropriate revision before building.

Once all the git checkout logic is moved into repo sync, Sheriffs will
be able to easily debug git flakiness failures -- if repo sync fails, we
know that it's a git problem, and we can rule out other types of build
problems.

This change should speed up our buildbots. I benchmarked this change on
the kernel alone and found it cut about 10 minutes off the build time.
For remote users, who might not have as fast a connection, this change
will lead to a bigger improvement in build times for a fresh chroot.

BUG=chromium-os:6454
TEST=Delete all temporary files to ensure local copies of git
     repositories aren't being used. Run setup_board --nousepkg &&
     build_packages --nousepkg both with and without the change to
     compare build times.

I got 56.75 minutes without the change, and 42.75 minutes with the change. Most of the speedup is for the kernel (10 minutes there alone). Neither of these numbers include icedtea.

Review URL: http://codereview.chromium.org/3312012

Change-Id: Ib4322344f59f324296a6682ba9056724c5e892f9