consolekit: add support for mocking a response for namespace pid'd objects
As part of the effort to clean up logout by parallelizing teardown,
VFS and PID namespacing is being added to the session_manager process
trees.  When a process starts a new dbus connections in the namespace, the
namespace'd pid is given to dbus and therefore consolekit.  This causes
ck-session-collect-info to fail. Its failure causes libcros's check
version code to fail and break incognito mode.

This change stubs out the call if ck-collect-session-info fails.  Since
we don't use consolekit as a security boundary, this should be okay.
Longer term, we can either fix the socket bug around this agl@ reported
in the kernel or stop using consolekit (or something else).

TEST=runs in a frankenbuild which uses namespace pids for incognito mode
BUG=chromium-os:6615

Change-Id: Ic9861d0f6ce9f2ddb7b1af7352bfe66dedf75dc6

Review URL: http://codereview.chromium.org/3357025