Make sure the recovery signing key makes it into the saft directory.
This CL is the ebuild modifying counterpart of
http://codereview.chromium.org/3347020/show.

There is a need to be able to re-sign the recovery kernel
blob when running SAFT. Which in turn requires that the
appropriate private development key is available on the
target.

This is achieved by making SAFT dependent on vboot_reference
and making vboot_reference ebuild place the proper key in
the SAFT directory.

Note that the saft tree does not have the x86-generic
directory anymore, the firmware ebuild is being changed
to copy files from one directory only.

Change-Id: Ic487c77ed64cdca2aba07d0decfe8b0d18ff021a

BUG=chromium-os:6345
TEST=manual

Run the following commands:

(chroot $) build_packages
(chroot $) build_images

install the resulting image on the target and observe the
proper key placed in /usr/sbin/firmware/saft.

Review URL: http://codereview.chromium.org/3297023