Remove the -nopie flag from the chromeos-chrome ebuild.
It seems that someone made Chrome work with the hardened compiler, as removing the -nopie flag works with no patching now.  Also it seems that the existing Chrome in the latest builds (0.8.70.2) seem to be PIE (looking at the output of readelf -h chrome, it is labeled as DYN).

Since this flag seems to have no effect anymore, and it seems we are building PIE anyway, I'd like to remove it from the ebuild.

Change-Id: Ic4b8cdbfe25342ce2f1f2c75379be0bb70b4abe0

BUG=4220
TEST=Build Chrome from source, install

Review URL: http://codereview.chromium.org/3343019