chromeos-base/kernel: move headers back into kernel-headers
kernel-headers are an implicit dependency of many packages. Having the
kernel headers in the kernel ebuilds causes potentional build breaks
due to the kernel ebuild being re-installed at the same time as a package
which depends on it is being installed. We're fixing this by moving
the headers back into kernel-headers.

kernel-headers will not be a cros-workon ebuild. It will use a pinned
GIT sha1.

I also modified setup_board to install kernel-headers instead of kernel.
I also modified build_packages to emerge the kernel before anything
else in order to avoid a collision between the old kernel and the new
kernel-headers ebuild.

BUG=4967
TEST=Did the following testing.

Verified that build_packages works for x86-generic and st1q on an old chroot.
Verified that the msm_* headers appear in /build/st1q but not in
/build/x86-generic.
For both cases above, verified that the upgrade from old kernel ebuid
(with headers) upgrades successfully to the new ebuild (without headers).

Change-Id: I2052239060d6107a540ef360ff359ab94b63e10e

Review URL: http://codereview.chromium.org/3257007